Fix: The program name wasn't being stripped when passed to exec
In the last commit, the way we parsed arguments was modified so that
the program name could be extracted. This introduced a bug that when
doing exec, the program name was given as a first argument.

The fix is to `skip` the first argument for the argument list given to
exec.