Use `eval -S` instead of arrays to build complex execution
Using `eval -S` allows us to use environment variables, in the form of
`${EDITOR}`, to decide which program to use. Not only that but it also
splits our input at the whitespace which makes the last commit that
allows us to take arrays no longer necessary.

With this change, you can now launch an editor in a terminal with the
simplest syntax possible:

```
"kitty ${EDITOR}"
```